requirements.txt
gunicorn
Flask
mysql-connector-python